pyproject.toml
[project]
name = "selenium-thsr"
version = "0.1.0"
description = ""
authors = [
    {name = "py-386",email = "billy121458596@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "selenium (>=4.28.1,<5.0.0)",
    "ddddocr (>=1.5.6,<2.0.0)",
    "certifi (>=2025.1.31,<2026.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
